FW: fix platform leaks in provisioning (F6)
- services: ProvisioningShell (no Arduino/Serial/esp_); handle_line() returns response + reboot flag
- platform: ProvisioningAdapter owns Serial I/O and esp_restart in .cpp only
- app: AppServices header forward-declares ProvisioningAdapter, wiring in .cpp
- remove services/serial_provisioning.h/.cpp
Refs #249
