+++ firmware/src/app/app_services.h
@@ -6,12 +6,15 @@
 #include "domain/logger.h"
 #include "services/oled_status.h"
 #include "services/radio_smoke_service.h"
-#include "services/serial_provisioning.h"
 
 namespace naviga {
 
+class ProvisioningAdapter;
+
 class AppServices {
  public:
+  AppServices();
+  ~AppServices();
   void init();
   void tick(uint32_t now_ms);
 
@@ -27,7 +30,7 @@ class AppServices {
   domain::Logger event_logger_;
   M1Runtime runtime_;
   OledStatus oled_;
-  SerialProvisioning serial_provisioning_;
+  ProvisioningAdapter* provisioning_ = nullptr;
 };
 
-} // namespace naviga
+}  // namespace naviga
