+++ firmware/src/platform/provisioning_adapter.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <cstddef>
+#include <cstdint>
+
+#include "services/provisioning_shell.h"
+
+namespace naviga {
+
+/**
+ * Platform adapter: UART I/O + non-blocking line reader; calls ProvisioningShell;
+ * performs reboot when shell requests it. Implementation uses Arduino/esp_* in .cpp only.
+ */
+class ProvisioningAdapter {
+ public:
+  /** Call once after E220 begin(); result_enum = E220BootConfigResult (0=Ok, 1=Repaired, 2=RepairFailed). */
+  void set_e220_boot_info(int result_enum, const char* message);
+
+  /** Read one line (non-blocking), handle via shell, print response; at most one line per call. */
+  void tick(uint32_t now_ms);
+
+ private:
+  ProvisioningShell shell_;
+  char line_buf_[ProvisioningShell::kLineMax] = {};
+  size_t line_len_ = 0;
+};
+
+}  // namespace naviga
